Roman numerals in Ruby
I used ruby's metaprograming to do this exercise. I added a "to_roman"
method inside ruby's internal "Fixnum" class.
This is not the most maintable solution, it just just for the purpose
of practicing one of the most powerfull tools of ruby - metaprograming.
In a real project I would avoid using it, it makes code hard to
understand. A better aproach would be to use a new type "Decimal" or create an
explicit convertor.

I like how readable it is in ruby to loop a collection with the "each"
notation rather than using a "while" or "for" in another language.

I came from doing the same exercise in python before this.
I've got a bit confused with the notation to declare fields in a class. Python uses
"self.variable" notation while ruby "@.variable" (for instance
variables) or "@@.variable" (for class variables). The self keyword is
used in both languages to refer the current object, that's why the
confusion.
I like more python when defining blocks with no need of the "end"
keyword - less verbose and concise code. Python uses ":" to open a block
and indentation to see when it should be closed.