Fizzbuzz in Erlang
I've already had done some Erlang before so not big deal.
I've used pattern matching for the algorithm. This makes the code more
readable than using if statements in one "big" method like we do with
no functional languages.

As part of the main fizzbuzz algorithm I created a way to print to
console the first N fizzbuzz numbers in the console. This bit took
longer than expected. Probably due to my lack of knowledge of how to
print to console, read arguments from the command line or work with
lists.
I found printing into the console way too complicated. It's kind of
similar to printf in java where you have to pass a string with the
format and then a list of variables. It's much easier in Ruby,
where we can use string interpolation.
I found also the syntax to interact with lists not the most readable or
"compact" - "lists:nths(index, List)" just to retrive an element or
"lists:foreach(Fun, List)" to loop a list.
Overall the documentation of erlang is good, but I struggled a bit too
find what I was looking for easily. I had to read loads of text and
there weren't many code samples. Also there are not as many websites in
google as you would find in languages like Ruby or C#.