Roman numerals in erlang
This one was much harder than expected. I have had done roman numerals
many times in C# and some other languages with multiple contraints, but
never with a functional language where I had to use recursion.

Erlang doesn't have loops so I couldn't use a while. In functional
language we will just call the same function recursively instead. Let's
say we want to iterate a collection, we can pass the collection as
argument to the function and use pattern matching to take the head
element, do something with it and call again to the same function with
the rest of the collection. Syntax for that would be [Head | Rest].
It took me a while to understand that's how I had to complete this
kata.

My first approach was to use just pattern matching and see what would
happen. After writting a few tests I ended up with the same function
duplicated all over the place but with different guard clauses. I didn't
know how to refactor this. I had to restart again and try it again, but
still couldn't make it work.. I had to finally watch the kata completed
in a youtube video to see how it was done.. I completed kind of copying
it, then next day I undid my changes and done it by myself. Yes, this is
cheating! But this way I learnt how we loop in functional languages. I
think I still have too much to learn in the functional world! :)