Fizzbuzz in elixir
I have used a "case" in one single method instead of having different
method overloads with different guards like I did it in erlang. I like
more this approach, but I don't think there is an approach better than the
other.

I came to do this kata from erlang, which is the "parent" of elixir.
Elixir is a new syntax on top of erlang, nicer to read. Quite similar to
ruby.

I liked how easy is to import functions from other modules. You just
call them and that's it! No need to explicitly declare using or exports
like you have to in other many languages. If you want to hide functions in
a module you can make them private.